refactor: rename components in /pages
