+++ src/App.tsx
@@ -3,10 +3,10 @@ import { Route, Routes } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from 'store/store';
 
-import { Home } from 'pages/home/Home';
-import { Cart } from 'pages/cart/Cart'; 
-import { ProductPage } from 'pages/productPage/ProductPage';
-import { Page404 } from 'pages/page404/Page404';
+import { Home } from 'pages/Home';
+import { Cart } from 'pages/Cart'; 
+import { ProductPage } from 'pages/ProductPage';
+import { Page404 } from 'pages/Page404';
 import { Footer } from 'layouts/Footer';
 import { Header } from 'layouts/Header';
