+++ src/pages/home/Home.test.tsx
@@ -1,4 +1,4 @@
-import { returnDisplayModeStyle } from "./Home";
+import { returnDisplayModeStyle } from "./index";
 
 test('returns correct display mode style', () => {
     expect(returnDisplayModeStyle('rows')).toBe('row');
